Refactor the message handling of the service worker to a separate message handler module
